+++ src/command/mod.rs
@@ -1,7 +1,7 @@
 mod parser;
 mod types;
 
-pub(crate) use types::{Command, CommandError, ProtocolVersion};
+pub(crate) use types::{ClientInfoAttribute, Command, CommandError, ProtocolVersion};
 
 #[cfg(test)]
 mod tests;